task/refactor/be - Extract .env/local.env reading logic from db.js #46
